requirements.txt
groq==0.9.0
python-dotenv==1.0.1
pandas==2.2.3
